requirements-full.txt
fastapi==0.85.0
uvicorn[standard]==0.20.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.12
pyjwt==2.6.0

# Async database dependencies
aiosqlite==0.17.0
asyncpg==0.28.0

# HTTP client for API integrations
httpx==0.24.0

# Environment variables
python-dotenv==0.19.0

# AI/ML dependencies (minimal versions)
openai==0.28.0
anthropic==0.3.0
requests==2.31.0

# Executable Agent Dependencies
twilio==8.5.0
sendgrid==6.9.0